+++ server/app/lib/http-api-server.js
@@ -411,6 +411,7 @@ var HttpAPIServer = function () {
     value: function getServerMKLivestatus(req, res) {
       this.logger.debug('HTTP request received: Get last MKLivestatus state');
       res.json({
+        lastCheck: this.stationManager.lastMKLivestatusDumpTime,
         lastState: this.stationManager.lastMKLivestatusDump
       });
     }