+++ server/src/lib/mk-livestatus-connector.js
@@ -129,16 +129,16 @@ export default class MKLivestatusConnector {
           } else {
             // Current data
             const matches = station.app_id.match(/^[^:]+:\s*(.*)@\[.*\]$/);
-            if (matches !== null && ('length' in matches) && matches.length > 1) {
+            if (station.app_state === Nagios.ServiceState.OK &&
+              matches !== null && ('length' in matches) && matches.length > 1) {
               station.app_id = matches[1];
-            } else if (station.app_id === 'CRIT - CRITICAL - no running TOP app!') {
+            } else if (station.app_state === Nagios.ServiceState.CRITICAL &&
+              station.app_id === 'CRIT - CRITICAL - no running TOP app!') {
               // Not really a critical error
               // There's just no app running. It happens. No need to call the Avengers.
               station.app_id = '';
               station.app_state = Nagios.ServiceState.OK;
               station.app_state_type = Nagios.StateType.SOFT;
-            } else {
-              throw new Error(`Error parsing app_id of station ${station.id}: ${station.app_id}`);
             }
           }
         }